requirements.txt
Django>=4.2,<5.0
djangorestframework
django-filter>=2.4.0,<3.0
openpyxl
reportlab
